Show editor+board: Kopfstreifen folders with title, ms-precise cue/length text inputs, seekable timeline playhead (local+Spotify), panic tile on subpages, fixed-size board labels, labeled profile duplicate/delete, two-column wider dialog + no horizontal scroll, ESC to close, discard-confirm when no song chosen

+++ src/RedAnts.Show/Features/Show/ShowEditModels.cs
@@ -6,8 +6,8 @@ public sealed class EditSound
 {
     public SoundKind Kind { get; set; } = SoundKind.Local;
     public string Ref { get; set; } = "";
-    public int StartSec { get; set; }
-    public int? DurationSec { get; set; }
+    public double StartSec { get; set; }
+    public double? DurationSec { get; set; }
     public bool Shuffle { get; set; }
 
     public EditSound Clone() => new() { Kind = Kind, Ref = Ref, StartSec = StartSec, DurationSec = DurationSec, Shuffle = Shuffle };
